refactor main for organisation
